Notebooks for two viable water mass budget methods
Method 1: using the cumulative mask (as we've done before), which is
fixed in time and guarantees that the budget closes

Method 2: using a rolling 7-day cumulative mask, which should still give
consistent budgets but does not have the double-counting problem of the
fixed cumulative mask.

There may be better methods than these two, but that will be left
for future work.